Port to Solidus
This is a WIP. It has not been tested interactively yet. This is purely
just replacing `spree` with `solidus` where needed and making the automated
tests pass.

I removed the `taxable_spec.rb` completely as it has changed so much the
test would need to be rewritten. Also the previous tests didn't actually
test the method it claimed it was testing. So for now just removing that
complication.