Fix a broken caching strategy
Previously, only the tax amount for the last line item would be used to
calculate the taxes for an order.

For instance if I had an item with a tax of 0.5 and an item with a tax of
1.2. The total tax would have been 2 * 1.2.

Now, we cache the whole TaxJar response and properly pick the correct
tax amount for each item. The cache key for the TaxJar response should
make sure that any update to the items, address, or shipping will result
in a new cache key, and thus an new TaxJar call to fetch up-to-date
amounts.

Signed-off-by: Rémy Coutable <remy@rymai.me>