Fix an issue with the TaxjarCalculator#order_cache_key method
The previous key wouldn't take in account a new line item with the exact
same attributes as a former one. For instance, if you add an item to
your cart, then go through the ordering process (at least to the
shipping tax rates calculation), and then remove the item from the cart
entirely, and then re-add it (it will create a new `LineItem` object,
but we shouldn't reuse the cached Taxjar order response in that case
since the former line item doesn't exist in the cached response.

Signed-off-by: Rémy Coutable <remy@rymai.me>